What's new?
What's new in this update, 5th November 2024

I have created my own platform, called RTK+ (Ryze Tha Kidd Plus), where you can pay to download my songs in one tab and you can watch my visualisers, lyric videos, behind-the-scenes and official music videos. Behind-the-scenes and official music videos will be locked behind a paywall, but the lyric videos and visualisers will be free to watch on YouTube. Head over to the RTK+ tab to see the new songs ready to be purchased and downloaded straight to your device, NO MONTHLY SUBSCRIPTION; JUST A ONE-TIME PAYMENT.

You can even go to the discography and look at the single you want and go to it's page and click the "PURCHASE" button to buy it.

PRICES
- Singles = £1.29
- EP's = £1.29 multiplied by however many songs are in the EP.
- Albums = £1.29 multiplied however many songs are in the album. If the album has over, or equal to, 8 songs, there will be a discount of 35% added to the final price. I will take care of all of that so don't worry about needing to get a discount code from anywhere, I'll update the price as necessary. If you are buying an album with 22 songs, the original price would be: £28.38. But with this discount, it will be £18.45.